LayoutTests/http/tests/resourceLoadStatistics/log-delayed-client-side-redirects-expected.txt

Non-user initiated, delayed cross-site navigations done client-side should be detected as top frame redirects by ITP.

Resource load statistics:

Registrable domain: 127.0.0.1
    hadUserInteraction: No
    mostRecentUserInteraction: -1
    grandfathered: No
    TopFrameUniqueRedirectsTo:
        localhost
    IsScheduledForAllButCookieDataRemoval: No
    isPrevalentResource: No
    isVeryPrevalentResource: No
    dataRecordsRemoved: 0
Registrable domain: localhost
    hadUserInteraction: No
    mostRecentUserInteraction: -1
    grandfathered: No
    TopFrameUniqueRedirectsFrom:
        127.0.0.1
    IsScheduledForAllButCookieDataRemoval: No
    isPrevalentResource: No
    isVeryPrevalentResource: No
    dataRecordsRemoved: 0
